Move onion encryption/decryption/etc into an onion_utils module
